Fixes to build and play on VMS x86 without BLISS.
Add a MACRO-32 version of the variable-argument-list wrapper that was
originally written in MACRO-32 and then converted to BLISS, presumably
for portability. Unfortunately, the x86 version of the BLISS compiler
isn't ready yet, and it's just as efficient to have MACRO-32 generate
the argument list from the register params, especially since there can
be so many of them (up to 9 for ifc$message or 7 for ifc$message_indent).

Besides this change, there was a bug in reading objects, like the sign
in the first room, or the ASCII table in the second room, because the
call to ifc$message() had some extra "0, 0" parameters for some reason.
This caused it to print "The sign reads:" but then nothing after that.

With these diffs, the game should now build and play on OpenVMS V9.2-1.