Revert change to $message parameters.
My workaround to get messages printing by removing pairs of 0, 0 params
wasn't correct. The code is supposed to print a blank line when it sees those.
The reason it's failing on x86 is due to a VSI Pascal compiler bug with
default parameters causing bad behavior at the default optimization level.

I'm going to try to work around the bug by modifying function put_scroll_dx
to not use "[truncate]".