Workaround for Pascal x86 [truncate] code gen bug.
The VSI Pascal compiler on x86 is generating bad code when default params
are used. This was causing some game messages not to print because of
embedded blank lines (a pair of 0 params in the ifc$message call) causing
put_scroll_dx to be called with no arguments, which was stopping the
rest of the message from being printed.

I reproduced the bad behavior with a standalone Pascal test program that
checks the presence of its parameters and prints them if present. It
behaves correctly when compiled with "/NoOpt" but crashes with the
default optimization level. I've reported this issue on the VSI Forum.

As a workaround, commented out the "[truncate]" modifier on all three
parameters, and passed default params in the one location where it was
being called with no arguments (the case that was failing). This fixes
the game text.